requirements.txt
pytest==7.2.1
flake8==6.0.0
pyyaml==6.0
Faker==16.8.1